porting ResetServiceIT to the common testing suite
